cutout/worker/requirements.txt
pyyaml
requests
rpdb
numpy
pandas
Pillow
astropy
mpi4py
